Fixed Python support
Previously we failed to import the 'yaml' module (which is supplied by the external PyYAML package).
This happened even if PyYAML is installed to the global version of Python, because pre-commit-hooks uses a new virtualenv so it can't access the globally installed packages.
In order to fix this, we copied the PyYAML 3.13 files to our code (for both Python2 and Python3) and imported the copied versions from setup.py.